requirements.txt
APScheduler==3.11.0
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
PyYAML==6.0.2
TgCrypto==1.2.5
tzlocal==5.2
uvloop==0.21.0
